+++ app/components/picture-tool-control.tsx
@@ -32,14 +32,15 @@ export function PictureToolControl(props: ToolControlButtonProps) {
       {isActive && (
         <Portal className="fixed max-h-[80vh] left-20 shadow-lg inset-y-4 my-auto overflow-y-auto">
           <ul className="flex flex-col gap-2 border border-default bg-screen/80 p-2 rounded">
-            {pictureSprites.map(({ name, ...sprite }) => (
-              <li key={name}>
+            {pictureSprites.map(({ picture, ...sprite }) => (
+              <li key={picture.name}>
                 <button
                   className="inline-flex shrink-0 cursor-pointer items-center hover:bg-primary-hover/80 active:bg-primary-active/80 justify-center gap-2 rounded text-sm font-bold transition-colors h-12 w-12"
                   onClick={() => {
-                    setToolState<PictureToolState>(defaultTools.picture.id, {
-                      name,
-                    });
+                    setToolState<PictureToolState>(
+                      defaultTools.picture.id,
+                      picture
+                    );
                   }}
                 >
                   <PictureIcon className="w-full h-full" src={sprite.src} />